requirements.txt
absl-py==0.11.0
amqp==5.0.2
argon2-cffi==20.1.0
asgiref==3.3.1
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
billiard==3.6.3.0
bleach==3.2.1
Box2D==2.3.10
Brotli==1.0.9
cachetools==4.1.1
celery==5.0.5
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
chitra==0.0.20
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cloudpickle==1.6.0
colorama==0.4.4
cycler==0.10.0
dash==1.18.1
dash-core-components==1.14.1
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.1
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.3
dj-database-url==0.5.0
dj-email-url==1.0.1
Django==3.1.3
django-cache-url==3.1.2
django-crispy-forms==1.9.2
djangorestframework==3.12.2
dnspython==1.16.0
entrypoints==0.3
environs==8.0.0
eventlet==0.30.0
Flask==1.1.2
Flask-Compress==1.8.0
flask-ngrok==0.0.25
future==0.18.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googleapis-common-protos==1.52.0
greenlet==0.4.17
grpcio==1.33.2
gunicorn==19.9.0
gym==0.17.3
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-resources==3.3.0
ipykernel==5.3.4
ipynb-py-convert==0.4.6
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
kombu==5.0.2
Markdown==3.3.3
MarkupSafe==1.1.1
marshmallow==3.9.1
matplotlib==3.3.3
memory-profiler==0.58.0
mistune==0.8.4
mlxtend==0.17.3
mnist==0.2.2
mplot3d-dragger==0.0.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nbresuse==0.4.0
nest-asyncio==1.4.3
networkx==2.5
notebook==6.1.5
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.4
pandoc==1.0.2
pandocfilters==1.4.3
parso==0.7.1
pathlib==1.0.1
pickleshare==0.7.5
Pillow==8.0.1
plotly==4.14.1
ply==3.11
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.8
protobuf==3.14.0
psutil==5.7.3
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyglet==1.5.0
Pygments==2.7.2
pyparsing==2.4.7
pyrsistent==0.17.3
pyterminal==0.4
pytesseract==0.3.6
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
PyWavelets==1.1.1
pywin32==300
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==20.0.0
redis==3.5.3
requests==2.25.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.4
Send2Trash==1.5.0
six==1.15.0
SQLAlchemy==1.3.22
sqlite3-api==1.0.7
sqlparse==0.4.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-addons==0.11.2
tensorflow-cpu==2.3.1
tensorflow-datasets==4.1.0
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.1
tensorflow-gpu-estimator==2.3.0
tensorflow-metadata==0.25.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
tf-keras-vis==0.5.4
tfds-nightly==4.1.0.dev202011210106
threadpoolctl==2.1.0
tifffile==2020.11.18
tornado==6.1
tqdm==4.52.0
traitlets==5.0.5
typeguard==2.10.0
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.1.0
widgetsnbextension==3.5.1
wrapt==1.12.1
